+++ src/ReactStickyHeader.js
@@ -6,7 +6,6 @@ import React, { Component } from 'react';
 import cx from 'classnames';
 
 import { addEvent } from './utils';
-import './styles.less';
 
 type Props = {
   children?: Children,